Attempted Random Forest, KDN Machine Learning Models
Added Property to Mieliesearch

Migrated DB to mySQL as Dockerfile rebuild attempt to implement ElasticSearch irrevocably broke MS SQL

Updated Dusk Tests

+++ database/migrations/0001_01_01_000014_create_needs_table.php
@@ -84,9 +84,15 @@ return new class extends Migration
      */
     public function down(): void
     {
-        // Turn off System Versioning
-        DB::statement("ALTER TABLE needs SET (SYSTEM_VERSIONING = OFF)");
-        Schema::dropIfExists('needs');
-        Schema::dropIfExists('needs_history');
+
+        $db_type = DB::connection()->getDriverName();
+
+        if ($db_type === 'sqlsrv') {
+
+            // Turn off System Versioning
+            DB::statement("ALTER TABLE needs SET (SYSTEM_VERSIONING = OFF)");
+            Schema::dropIfExists('needs');
+            Schema::dropIfExists('needs_history');
+        }
     }
 };
